Refactoring to MemoryLimit object for QueueCommand (#72)
* MemoryLimit moved into eLife\Search\Gearman

* Added MemoryLimits for QueueCommand

* Removed duplicated --memory* options

* Coding standards

* Field for MemoryLimit
